# docs/swagger.yaml
basePath: /
definitions:
...
  hauslet_internal_modules_property_domain.DeleteMediaRequest:
    properties:
      media:
        items:
          $ref: '#/definitions/hauslet_internal_modules_property_domain.MediaDeleteInput'
        maxItems: 50
        minItems: 1
        type: array
    required:
    - media
    type: object
  hauslet_internal_modules_property_domain.ErrorResponse:
    properties:
      error:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  hauslet_internal_modules_property_domain.FinalizeMediaRequest:
    properties:
      media_keys:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - media_keys
    type: object
  hauslet_internal_modules_property_domain.MediaDeleteInput:
    properties:
      key:
        type: string
      media_id:
        type: string
    required:
    - key
    - media_id
    type: object
  hauslet_internal_modules_property_domain.MediaType:
    enum:
    - image
    - video
    - 360_tour
    - document
    type: string
    x-enum-varnames:
    - MediaTypeImage
    - MediaTypeVideo
    - MediaTypeTour360
    - MediaTypeDocument
  hauslet_internal_modules_property_domain.MediaUploadInput:
    properties:
      caption:
        type: string
      duration:
        type: integer
      filename:
        type: string
      group:
        type: string
      is_group_cover:
        type: boolean
      is_primary:
        type: boolean
      mime_type:
        type: string
      order:
        type: integer
      size_bytes:
        type: integer
      type:
        $ref: '#/definitions/hauslet_internal_modules_property_domain.MediaType'
    required:
    - filename
    - size_bytes
    - type
    type: object
  hauslet_internal_modules_property_domain.MediaUploadResult:
    properties:
      filename:
        type: string
      id:
        type: string
      key:
        type: string
      url:
        type: string
    type: object
  hauslet_internal_modules_property_domain.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  hauslet_internal_modules_property_domain.UpdateMediaRequest:
    properties:
      caption:
        type: string
      is_group_cover:
        type: boolean
      is_primary:
        type: boolean
      order:
        type: integer
    type: object
  hauslet_internal_modules_property_domain.UploadMediaRequest:
    properties:
      media:
        items:
          $ref: '#/definitions/hauslet_internal_modules_property_domain.MediaUploadInput'
        maxItems: 50
        minItems: 1
        type: array
    required:
    - media
    type: object
  hauslet_internal_modules_property_domain.UploadMediaResponse:
    properties:
      media:
        items:
          $ref: '#/definitions/hauslet_internal_modules_property_domain.MediaUploadResult'
        type: array
      total:
        type: integer
    type: object
...
host: https://dev-api.hauslet.com
info:
  contact:
    email: support@hauslet.com
    name: API Support
    url: http://www.swagger.io/support
  description: Hauslet Services API Server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Hauslet API
  version: "1.0"
paths:
...
  /listings/{id}/media:
    delete:
      consumes:
      - application/json
      description: Delete one or more media items from a listing
      parameters:
      - description: Listing ID
        in: path
        name: id
        required: true
        type: string
      - description: Media delete details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hauslet_internal_modules_property_domain.DeleteMediaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete listing media
      tags:
      - listings
    post:
      consumes:
      - application/json
      description: Generate presigned upload URLs for listing media
      parameters:
      - description: Listing ID
        in: path
        name: id
        required: true
        type: string
      - description: Media upload details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hauslet_internal_modules_property_domain.UploadMediaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.UploadMediaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload listing media
      tags:
      - listings
  /listings/{id}/media/{mediaId}:
    patch:
      consumes:
      - application/json
      description: Update metadata for a specific media item
      parameters:
      - description: Listing ID
        in: path
        name: id
        required: true
        type: string
      - description: Media ID
        in: path
        name: mediaId
        required: true
        type: string
      - description: Media update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hauslet_internal_modules_property_domain.UpdateMediaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update listing media
      tags:
      - listings
  /listings/{id}/media/finalize:
    post:
      consumes:
      - application/json
      description: Mark media as uploaded and trigger thumbnail generation
      parameters:
      - description: Listing ID
        in: path
        name: id
        required: true
        type: string
      - description: Media finalization details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hauslet_internal_modules_property_domain.FinalizeMediaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_property_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Finalize listing media
      tags:
      - listings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"